pyproject.toml
[project]
name = "fg-forge-updater"
version = "1.4.0"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "bmos", email = "wil.thieme@protonmail.com" }
]
urls = { Repository = "https://github.com/bmos/fg_forge_updater" }
dependencies = [
    "beautifulsoup4==4.13.3",
    "markdown==3.7",
    "matplotlib==3.10.1",
    "mdformat-footnote==0.1.1",
    "mdformat-frontmatter==2.0.8",
    "mdformat-gfm==0.4.1",
    "pandas==2.2.3",
    "python-dotenv==1.1.0",
    "requestium==0.4.0",
    "seaborn==0.13.2",
    "selenium==4.30.0",
]

[project.optional-dependencies]
dev = [
    "bandit[toml]==1.8.3",
    "mypy==1.15.0",
    "pandas-stubs==2.2.3.250308",
    "polyfactory==2.19.0",
    "pre-commit==4.2.0",
    "pytest-xdist==3.6.1",
    "pytest==8.3.5",
    "ruff==0.11.4",
    "types-Markdown==3.7.0.20241204",
    "types-beautifulsoup4==4.12.0.20250204",
    "types-requests==2.32.0.20250301",
    "types-seaborn==0.13.2.20250111",
]
github-actions = [
    "coverage==7.8.0",
    "pytest-cov==6.0.0",
]

[tool.ruff]
line-length = 160
target-version = "py311"
include = [
    "src/**/*.py",
    "tests/**/*.py",
]

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ARG",
    "B",
    "BLE",
    "C",
    "C4",
    "C90",
    "D",
    "DTZ",
    "E",
    "EM",
    "F",
    "G",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "N",
    "PT",
    "PTH",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "T20",
    "TC",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "D100",
    "D103",
    "D104",
    "D203",
    "D212",
]

[[tool.mypy.overrides]]
module = [
    "requestium",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=src -n auto"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
relative_files = true
command_line = "-m pytest"

[tool.coverage.paths]
source = [
    "src/",
]
omit = [
    "tests",
]

[tool.coverage.report]
exclude_also = [
    "logger.",
]

[tool.bandit]
exclude_dirs = [
    "tests/",
    ".venv/",
]
